+++ international/templates/international/footer.html
@@ -18,8 +18,7 @@
                     <li class="govuk-footer__inline-list-item">
                         <a class="govuk-link govuk-link--inverse"
                            id="footer-contact"
-                           href="{% url 'international_online_offer:contact' %}?next={{ feedback_next_url }}">Contact us
-                        </a>
+                           href="{% url 'international:contact' %}?next={{ feedback_next_url }}">Contact us</a>
                     </li>
                     <li class="govuk-footer__inline-list-item">
                         <a class="govuk-link govuk-link--inverse"
