+++ international/templates/international/includes/form_field.html
@@ -1,0 +1,20 @@
+{% if field %}
+    <label class="govuk-label govuk-label--s"
+           for="{% if labelForId %}{{ labelForId }}{% else %}{{ field.id_for_label }}{% endif %}">
+        {{ field.label }}
+        {% if optional %}<span>(Optional)</span>{% endif %}
+        {% if dynamicText %}{{ dynamicText }}{% endif %}
+    </label>
+    {% if field.help_text %}<div class="govuk-hint govuk-!-margin-bottom-2">{{ field.help_text|safe }}</div>{% endif %}
+    {% for error in field.errors %}
+        <p class="govuk-error-message" role="alert">
+            <span class="govuk-visually-hidden">Error:</span> {{ error }}
+        </p>
+    {% endfor %}
+    {% if inputWrapperClass %}
+        <div class="{{ inputWrapperClass }}">
+            {% if inputPrefixClass %}<div class="{{ inputPrefixClass }}" aria-hidden="true">{{ inputPrefixText }}</div>{% endif %}
+        {% endif %}
+        {{ field }}
+        {% if inputWrapperClass %}</div>{% endif %}
+{% endif %}
