+++ international/templates/international/includes/submit_button.html
@@ -1,0 +1,1 @@
+<button type="submit" class="govuk-button" data-module="govuk-button">{{ button_text }}</button>
