+++ international/urls.py
@@ -1,3 +1,13 @@
+from django.urls import path
+
+from international import views
+
 app_name = 'international'
 
-urlpatterns = []
+urlpatterns = [
+    path(
+        'contact/',
+        views.ContactView.as_view(),
+        name='contact',
+    ),
+]
