+++ international_online_offer/views.py
@@ -1,5 +1,4 @@
 from directory_forms_api_client import actions
-from directory_forms_api_client.helpers import Sender
 from django.core.paginator import Paginator
 from django.http import HttpResponseRedirect
 from django.shortcuts import redirect, render
@@ -712,61 +711,6 @@ class FeedbackView(GA360Mixin, FormView):
             subject=self.subject,
             email_address='anonymous-user@expand-your-business.trade.gov.uk',
             form_url=self.request.get_full_path(),
-        )
-
-        response = action.save(cleaned_data)
-        response.raise_for_status()
-
-    def form_valid(self, form):
-        self.submit_feedback(form)
-        return super().form_valid(form)
-
-
-class ContactView(GA360Mixin, FormView):
-    form_class = forms.ContactForm
-    template_name = 'eyb/contact.html'
-    subject = 'EYB Contact form'
-
-    def __init__(self):
-        super().__init__()
-        self.set_ga360_payload(
-            page_id='Contact',
-            business_unit='ExpandYourBusiness',
-            site_section='contact',
-        )
-
-    def get_back_url(self):
-        back_url = '/international/expand-your-business-in-the-uk/guide/'
-        if self.request.GET.get('next'):
-            back_url = check_url_host_is_safelisted(self.request)
-        return back_url
-
-    def get_success_url(self):
-        success_url = reverse_lazy('international_online_offer:contact') + '?success=true'
-        if self.request.GET.get('next'):
-            success_url = success_url + '&next=' + check_url_host_is_safelisted(self.request)
-        return success_url
-
-    def get_context_data(self, **kwargs):
-        return super().get_context_data(**kwargs, back_url=self.get_back_url())
-
-    def submit_feedback(self, form):
-        cleaned_data = form.cleaned_data
-        if self.request.GET.get('next'):
-            cleaned_data['from_url'] = check_url_host_is_safelisted(self.request)
-
-        sender = Sender(
-            email_address=cleaned_data['email'],
-            country_code=None,
-        )
-
-        action = actions.ZendeskAction(
-            full_name=cleaned_data['full_name'],
-            email_address=cleaned_data['email'],
-            subject=self.subject,
-            service_name='expand your business',
-            form_url=self.request.get_full_path(),
-            sender=sender,
         )
 
         response = action.save(cleaned_data)
